snorkel_flow_backend/requirements.txt
absl-py==2.1.0
asgiref==3.7.2
click==8.1.7
Django==3.2.25
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-filter==23.2
django-zen-queries==2.1.0
djangorestframework==3.14.0
filelock==3.13.1
fsspec==2024.2.0
grpcio==1.62.1
Jinja2==3.1.3
joblib==1.3.2
Markdown==3.5.2
MarkupSafe==2.1.5
mpmath==1.3.0
munkres==1.1.4
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
pandas==2.2.1
Pillow==8.4.0
protobuf==4.25.3
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2023.12.25
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
snorkel==0.9.9
sqlparse==0.4.4
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
textblob==0.18.0.post0
the-social-network==0.0.7
threadpoolctl==3.3.0
torch==2.2.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
Werkzeug==3.0.1